requirements.txt
Flask>=0.12.1
Flask-Bcrypt>=0.7.1
Flask-WTF>=0.14.2
itsdangerous>=0.24
Jinja2>=2.9.6
Markdown>=2.6.8
MarkupSafe>=1.0
msgpack-python>=0.4.8
openpyxl>=2.4.5
peewee>=2.9.2
python-dateutil>=2.6.0
requests>=2.13.0
six>=1.10.0
virtualenv>=15.1.0
Werkzeug>=0.12.1